requirements.txt
requests
ipykernel
python-dotenv
openai
pydantic
docling
lancedb
streamlit
tiktoken
azure-identity
azure-keyvault-secrets